+++ ms-property-catalog/src/main/java/com/eduardo/mspropertiescatalog/repository/PropertyRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -14,5 +15,5 @@ public interface PropertyRepository extends JpaRepository<Property, UUID> {
     Optional<Property> findByAddress(String address);
     Page<Property> findByCity(ECity city, Pageable pageable);
 
-    Page<Property> findBypricePerNightBetween(Double minPrice, Double maxPrice, Pageable pageable);
+    Page<Property> findBypricePerNightBetween(BigDecimal minPrice, BigDecimal maxPrice, Pageable pageable);
 }
