+++ ms-reservation/src/main/java/com/eduardo/msreservation/client/dto/PropertyInfoDto.java
@@ -1,7 +1,9 @@
 package com.eduardo.msreservation.client.dto;
 
+import java.math.BigDecimal;
+
 public record PropertyInfoDto(
         String propertyId,
-        Double pricePerNight
+        BigDecimal pricePerNight
 ) {
 }
